feat: Increase coverage of the env and env-file options for global, group and target scope (#39)
This PR aims to improve the usage of `env` and `env-file` in order to be used in global, group, and target scope.

Variables defined in `env` section will override any variable defined by the shell environment  or defined in the .env file

This PR also propose to change the access of variables via `vars` and environment variables via `env`, so it could can be access as following, for example: `{{ vars.myvar }}` and `{{ env.myenv  }}`

Another important change here is the improvement in the outputs, so it would be able to be displayed to the CI output. And the Error Code now is handled by MakimError enum class. So for each kind of error, it raises a specific error code.